# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grants Explorer API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GrantItemExtra = { [key: string]: unknown };

export interface GrantItem {
  id: string;
  title: string;
  description?: string;
  amount?: string;
  deadline?: string;
  agency?: string;
  url?: string;
  status?: string;
  extra?: GrantItemExtra;
}

export interface GrantsGovResult {
  source: string;
  total?: number;
  items: GrantItem[];
}

export interface SbirResult {
  source: string;
  total?: number;
  items: GrantItem[];
}

export interface ThreeSixtyGivingResult {
  source: string;
  total?: number;
  items: GrantItem[];
}

export interface CaGrantsResult {
  source: string;
  total?: number;
  items: GrantItem[];
}

export interface UsaSpendingResult {
  source: string;
  total?: number;
  items: GrantItem[];
}

export interface NihResult {
  source: string;
  total?: number;
  items: GrantItem[];
}

export interface NsfResult {
  source: string;
  total?: number;
  items: GrantItem[];
}

export interface SimplerGrantsResult {
  source: string;
  total?: number;
  items: GrantItem[];
}

export interface WorldBankResult {
  source: string;
  total?: number;
  items: GrantItem[];
}

export type GetGrantsGovParams = {
  keyword?: string;
  rows?: number;
};

export type GetSbirParams = {
  keyword?: string;
  rows?: number;
};

export type GetThreeSixtyGivingParams = {
  keyword?: string;
  rows?: number;
};

export type GetCaGrantsParams = {
  keyword?: string;
  rows?: number;
};

export type GetUsaSpendingParams = {
  keyword?: string;
  rows?: number;
};

export type GetNihParams = {
  keyword?: string;
  rows?: number;
};

export type GetNsfParams = {
  keyword?: string;
  rows?: number;
};

export type GetSimplerGrantsParams = {
  keyword?: string;
  rows?: number;
};

export type GetWorldBankParams = {
  keyword?: string;
  rows?: number;
};