# lib/api-zod/src/generated/types/simplerGrantsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grants Explorer API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GrantItem } from "./grantItem";

export interface SimplerGrantsResult {
  source: string;
  total?: number;
  items: GrantItem[];
}